moved english controlelrs methods to concern
