+++ app/controllers/english_offers_controller.rb
@@ -1,12 +1,13 @@
-# rubocop:disable Metrics
+# frozen_string_literal: true
+
 class EnglishOffersController < ApplicationController
   include BeforeRender
   include Offerable
   protect_from_forgery with: :null_session
 
+  # order is important
   before_action :set_offer, only: %i[show edit update]
-  before_render :find_or_initialize_autobidder, only: %i[new create edit update]
-  before_action :prevent_check_for_invalid_bid, only: [:update]
+  include EnglishOffers::Offerable
 
   include RecaptchaValidatable
   recaptcha_action 'english_offer'
@@ -22,40 +23,25 @@ class EnglishOffersController < ApplicationController
 
   # POST /auctions/aa450f1a-45e2-4f22-b2c3-f5f46b5f906b/offers
   def create
-    unless check_first_bid_for_english_auction(create_params, @auction)
-      formatted_starting_price = format('%.2f', @auction.starting_price)
-      flash[:alert] = t('english_offers.create.bid_must_be', minimum: formatted_starting_price)
-
-      if turbo_frame_request?
-        render turbo_stream: turbo_stream.action(:redirect, root_path)
-      else
-        redirect_to root_path, status: :see_other and return
-      end
-    end
+    inform_about_invalid_bid_amount and return unless check_first_bid_for_english_auction(create_params, @auction)
+    inform_invalid_captcha and return unless recaptcha_valid
 
     @offer = Offer.new(create_params)
     @offer.username = Username::GenerateUsernameService.new.call
     authorize! :manage, @offer
 
-    if recaptcha_valid
-      if create_predicate(@auction)
-        broadcast_update_auction_offer(@auction)
-        send_outbided_notification(auction: @auction, offer: @offer, flash:)
-        update_auction_values(@auction, t('english_offers.create.created'))
-      else
-        flash[:alert] = if @offer.errors.full_messages_for(:cents).present?
-                          @offer.errors.full_messages_for(:cents).join
-                        else
-                          @offer.errors.full_messages.join('; ')
-                        end
-
-        redirect_to root_path and return
-      end
+    if create_predicate(@auction)
+      broadcast_update_auction_offer(@auction)
+      send_outbided_notification(auction: @auction, offer: @offer, flash:)
+      update_auction_values(@auction, t('english_offers.create.created'))
     else
-      @show_checkbox_recaptcha = true unless @success
-      flash.now[:alert] = t('english_offers.form.captcha_verification')
+      flash[:alert] = if @offer.errors.full_messages_for(:cents).present?
+                        @offer.errors.full_messages_for(:cents).join
+                      else
+                        @offer.errors.full_messages.join('; ')
+                      end
 
-      redirect_to root_path, status: :see_other and return
+      redirect_to root_path and return
     end
   end
 
@@ -74,71 +60,24 @@ class EnglishOffersController < ApplicationController
   def update
     @auction = Auction.english.with_user_offers(current_user.id).find_by(uuid: @offer.auction.uuid)
     redirect_to auction_path(@auction.uuid) and return if update_not_allowed(@auction)
+    inform_invalid_captcha and return unless recaptcha_valid
 
-    if recaptcha_valid
-      if update_predicate(@auction)
-        broadcast_update_auction_offer(@auction)
-        send_outbided_notification(auction: @auction, offer: @offer, flash:)
-        update_auction_values(@auction, t('english_offers.edit.bid_updated'))
-      else
-        flash[:alert] = if @offer.errors.full_messages_for(:cents).present?
-                          @offer.errors.full_messages_for(:cents).join
-                        else
-                          @offer.errors.full_messages.join('; ')
-                        end
-
-        redirect_to root_path
-      end
+    if update_predicate(@auction)
+      broadcast_update_auction_offer(@auction)
+      send_outbided_notification(auction: @auction, offer: @offer, flash:)
+      update_auction_values(@auction, t('english_offers.edit.bid_updated'))
     else
-      @show_checkbox_recaptcha = true unless @success
-      flash.now[:alert] = t('english_offers.form.captcha_verification')
-      redirect_to root_path, status: :see_other
+      flash[:alert] = if @offer.errors.full_messages_for(:cents).present?
+                        @offer.errors.full_messages_for(:cents).join
+                      else
+                        @offer.errors.full_messages.join('; ')
+                      end
+
+      redirect_to root_path
     end
   end
 
   private
-
-  def find_or_initialize_autobidder
-    @autobider = current_user&.autobiders&.find_or_initialize_by(domain_name: @auction.domain_name)
-  end
-
-  def broadcast_update_auction_offer(auction)
-    Offers::UpdateBroadcastService.call({ auction: })
-  end
-
-  def update_auction_values(auction, message_text)
-    AutobiderService.autobid(auction)
-    auction.update_ends_at(@offer)
-
-    flash[:notice] = message_text
-    redirect_to root_path
-  end
-
-  def prevent_check_for_invalid_bid
-    auction = Auction.with_user_offers(current_user.id).find_by(uuid: @offer.auction.uuid)
-
-    return unless bid_is_bad?(auction:, update_params:)
-
-    flash[:alert] =
-      "#{t('english_offers.show.bid_failed', price: format('%.2f', auction.highest_price.to_f).tr('.', ','))}"
-
-    if turbo_frame_request?
-      render turbo_stream: turbo_stream.action(:redirect, root_path)
-    else
-      redirect_to root_path, status: :see_other
-    end
-  end
-
-  def bid_is_bad?(auction:, update_params:)
-    !additional_check_for_bids(auction, update_params[:price]) ||
-      !check_bids_for_english_auction(update_params, auction)
-  end
-
-  def additional_check_for_bids(auction, current_bid)
-    order = auction.offers.order(updated_at: :desc).first
-
-    Money.new(order.cents).to_f < current_bid.to_f
-  end
 
   def create_predicate(auction)
     @offer.save && auction.update_minimum_bid_step(create_params[:price].to_f) && @offer.reload
@@ -146,24 +85,6 @@ class EnglishOffersController < ApplicationController
 
   def create_params
     params.require(:offer).permit(:auction_id, :user_id, :price, :billing_profile_id, :username)
-  end
-
-  def check_first_bid_for_english_auction(params, auction)
-    return true if auction.blind?
-
-    starting_price = auction.starting_price
-    price = params[:price].to_f
-
-    price.to_f >= starting_price.to_f
-  end
-
-  def check_bids_for_english_auction(params, auction)
-    return true if auction.blind?
-
-    minimum = auction.min_bids_step.to_f
-    price = params[:price].to_f
-
-    price >= minimum
   end
 
   def update_predicate(auction)
